Replace join aligner with new version, remove duplicate
