+++ src/main/java/io/github/mzmine/gui/mainwindow/MainMenu.fxml
@@ -379,8 +379,6 @@
     <Menu text="Alignment">
       <MenuItem text="Join aligner" onAction="#runModule"
         userData="io.github.mzmine.modules.dataprocessing.align_join.JoinAlignerModule"/>
-      <MenuItem text="Join aligner2" onAction="#runModule"
-        userData="io.github.mzmine.modules.dataprocessing.align_join.JoinAligner2Module"/>
       <MenuItem text="Merge lists" onAction="#runModule"
         userData="io.github.mzmine.modules.dataprocessing.align_append_rows.MergeAlignerModule"/>
       <MenuItem text="RANSAC aligner" onAction="#runModule"
